Time: 0 ms (100.00%) | Memory: 9.6 MB (50.75%) - LeetSync

+++ 2395-longest-binary-subsequence-less-than-or-equal-to-k/longest-binary-subsequence-less-than-or-equal-to-k.cpp
@@ -10,7 +10,7 @@ public:
             if(s[i]=='0')
             count++;
             else if(pow<32 && ans+(1LL<<pow)<=k){
-                ans=ans+(1LL<<pow);
+                ans=ans+(long long)(1<<pow);
             count++;
             }
             pow++;
